Fix a bug with GridLayout horizontal groups
Plus some cleanup and debugs

Occasionally, the isHoriz var could be nil (usually via spec change) and
since there was not a nil check in place this setting would get passed
to Grid2.

This would cause the Grid2LayoutFrame to be sized as if there were just
one single group arrayed horizontally. The units would still be properly
grouped, but are anchored to the bottom left of the layout making it
appear as if the raid frames are steadily moving left of center.

This is because Grid2 in many places just does a simple boolean check
and assumes if horizontal evals to false (ie. nil), then the user must
want vertical.