Fix a bug preventing proper updates for specialized trackers
The UNIT_AURA update will clear the count text if the aura is not present,
but this was interfering with the specialized trackers because the update
event (eg. CLEU) would fire before UNIT_AURA. This interaction would
cause the count to get cleared until the next event update, which might
also be followed by another UNIT_AURA.

To solve this, all tracker modifications have been moved to a postUnitAura
function that is created on first run of the specialized update.