Add a throttle to LibTextDump for large buffers
The `MAJOR` for the lib has been changed so other addons are not impacted
by this WIP feature. Once I get it in better shape I'll submit a patch.

With this change, if a given buffer is over a threshold (determined by
average line length and number of lines) a coroutine will be created. This
coroutine is used to throttle the `table.concat` operation to prevent a
client crash.