Redesign the previously implemented LibTextDump throttle
This design uses a FauxScrollFrame to limit the total size of the displayed
string, rather than coroutines to keep the game from crashing when
concatenating a very large buffer.

The immediate benefit of this design is near instant display of the dump
frame and text, regardless of the buffer size. The down side though is
you can no longer copy/paste the entire buffer since the text only ever
contains a subset of it.

To mitigate this down side, I have modified lib:New() and the highlight
button to allow for the host addon to save the associated buffer to its
saved vars.