Make dump frame scroll smoother
This is a fairly significant rewrite of the scrolling code, and probably
closer to what I should have done to begin with.

The previous implementation had significant issues when near or at the
bottom of the scroll, resulting in needing to scroll up then back down to
actually reach the last line. This was due to an incongruity between the
number of lines in the buffer, and the number of lines needing to be
displayed after line wrapping.

To overcome this, the entire buffer is now iterated over to see how many
lines the each entry will require. The total of this is used instead of
the buffer length to determine the scroll height.

The current line offset was also changed to now use a formula that scales
a given value within a range to a proportional place within a second given
range.