Refactor text update
Previously, the dump frame would appear to have empty lines. This was due to the fact that when scrolling, the position was based on the current size of the frame's buffer, which may have changed since the frame was opened.

This commit changes the update behavior to also add in new lines as they come in when a frame is open. This eliminate those empty lines, and provides context for possible scroll jittering.