Prevent AlertFrames from drifting
Loot roll frames get added to the GroupLootContainer, but since they had all events unregistered, they don't get removed. This causes the alert frames to "drift" due to GroupLootContainer's larger size.

There is likely a better solution to this, but it probably requires rewriting significant portions of this mod, which I'm not really up to do at the moment.