Fix scroll being reset
Using and/or short-circuiting in this
context was not a good idea since
frame:UpdateText() does not return anything.

This meant that self:Display() would get run
anyway, which would overwrite the old
UpdateText method and reset the local offset
to 1, making the viewed text reset to the
top of the buffer.