Add workaround for artifact data
Artifact info isn't available until after
ARTIFACT_UPDATE, so until then we fake it
by just returning true regardless.

However, the event won't fire if you don't
have an artifact. Since there isn't a
reliable way to know if a char has an
artifact early on, we just check if they are
above 100.

If they are we wait for accurate info, If not
though, we just use the API as is, inaccuracies
be damned. Shouldn't be a problem though since
an active char with an artifact will get to 101
pretty quickly.