Refactor cooldown count
This is a re-write of the text scaling code that keeps the cooldown text
within the frame. This was prompted primarily from certain encounter
mechanics in Nighthold which increase and decrease the duration of
cooldowns and auras.

When a cooldown is increased, the scale of the text would be reduced to
half but would not return to normal when the cooldown speeds up again.