Account for hidden blocks when updating positions
When a block is removed, the blocks after it get shifted down to fill the space. This is needed in order to maintain the dock, but what would happen is that this new position would be maintained even if the block was only temporarily hidden like for a mail notification. This meant that on the next load, the hidden block would get "pushed" by each successive block that gets added resulting in the hidden block being in the wrong place.

This commit fixes that by keeping track of what blocks are currently being hidden and adjusting the saved index for any subsequent blocks.