Refactor AbbreviateName
This refactor is doing a few things.

For some people, the UTF8 library functions are not available via the
string literal metatable. To workaround this, we are converting the
UTF8 functions to use dot syntax instead.

Second, we are removing gsub in favor of strsplit and a while loop.
This allows for more control over what words get truncated and how.