Fix an error when switching equip sets
If an equip set was deleted that had been chosen for a spec, and then
attempted to change the equip set for that spec an error would occur.

This is because the saved variable did not get updated when the set was
deleted. As a result, when looking for the next index it would trip
over the missing set.